Move to ec2 package
