+++ aws/ec2/ec2.go
@@ -1,4 +1,4 @@
-package aws
+package ec2
 
 import (
 	"encoding/csv"
